refactor(components): [autocomplete] refactor autocomplete (#6067)
